+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added `Data().add_data_frame_index(data_frame, name)`, with this `pandas.DataFrame.index` can be added to `Data()`.
+
 ## [0.8.1] - 2022-04-14
 
 ### Fixed